cosmos/ingestion/requirements.txt
click
pymongo
pandas
pdfminer.six
chardet
pillow
joblib
falcon
click
mysqlclient
sqlalchemy
gunicorn
requests
celery==4.2.1
redis==3.2.0
gevent
pillow
falcon
mysqlclient
sqlalchemy
gunicorn
pytest
torch
torchvision
numpy==1.15.4 
pyyaml 
hyperyaml
scikit-image
bitmath
tensorboardX 
tensorboard 
tqdm 
opencv-python 
click
pascal-voc-writer 
imgaug
sqlalchemy
dataclasses
matplotlib
pandas
xgboost==0.90
sklearn
bs4
pytesseract
flower
setuptools

